requirements.txt
psd_tools==1.8.22
pillow
numpy
opencv-python
scipy==1.3.0
